Introduced Pokerwars.Score
Instead of returning an atom, the score function
will now return a struct representing the score.

This will be used to provide more details about the score,
for example what three of a kind and pair form a specific
full house. This extra information will be used to
implement tie breaking rules.